Rebase fibers on top of schedulers and tasks
This commit refactors fibers to be based on lighter-weight "tasks",
and makes the scheduler API more orthogonal.  Now there are no more
fiber objects (although the fibers layer could re-add them if they are
useful), and fibers no longer have names (although again this could be
re-added at an upper layer).  Also it's the current scheduler that's
bound by a parameter, not the current fiber, and as it's a thread-safe
parameter it doesn't need to be bound in each fiber.