'schedule-task-when-fd-active' always adds FD finalizer.
Previously, if we repeatedly did, say:

  socket() = 12;
  schedule-task-when-fd-active(12);
  close(12);

then 'fd-waiters' would only be cleared the first time.  On subsequent
iterations, the 'fd-waiters' list would not be cleared, thus this
condition would be verified:

  (and active-events
       (= (logand events active-events) events))

This, in turn, means that the FD would not be re-added to the epoll set,
and events on that FD would never be served.

* fibers/scheduler.scm (schedule-task-when-fd-active)[fd-finalizer]: New
procedure.
Call 'add-fdes-finalizer!' before the first 'epoll-add*!' call.
Use 'fd-finalizer' instead of the local 'finalize-fd' procedure.
* tests/ports.scm: New file.
* Makefile.am (TESTS): Add it.