scheduler: Do not re-add FD finalisers on FDs that already have one.
Fixes #65.

Commit c25dcb9cc4b5b977474ffe555b40ce2f1d0d1edc introduced a memory
leak, whereby 'schedule-tasks-for-active-fd' would keep adding fdes
finalizers on actively-used file descriptors, for example when calling
'accept' several times on the same file descriptor.

This commit fixes it by allowing 'schedule-tasks-for-active-fd' to
distinguish between "finalized file descriptor" and "file descriptor
without waiters but with a finalizer already".

* fibers/scheduler.scm (schedule-tasks-for-active-fd): Set the car of
EVENTS+WAITERS to 0, not #f.
(schedule-task-when-fd-active): Treat the cases where FD-WAITERS = #f
and FD-WAITERS = '(#f) in the same way.  In the other case, assume
ACTIVE-EVENTS is an integer.