Define variant of dynamic-wind* interacting well with scheduling.
No more 'dynamic-wind is convenient, but it doesn't interact well with
scheduling'!.  Now you can have both userspace scheduling and
'dynamic-wind'!.

The idea is based on discussion on
<https://logs.guix.gnu.org/guile/2022-11-20.log>; Guile-Parallel has
more-or-less the same implementation.

* fibers/scheduler.scm (rewinding-for-scheduling?): New parameter.
(%nesting-test-1,%nesting-test-2?): New variables for tests.
(dynamic-wind*): New procedure.
(suspend-current-task): Set rewinding-for-scheduling? appropriately.
(yield-current-task): Likewise, and When %nesting-test-1 and not
nested?, invoke yield-current-task for testing. Likewise for %nesting-test-2.
* fibers.scm (dynamic-wind*): Re-export for convenience.
* tests/basic.scm: Add some tests for 'dynamic-wind*'.
* NEWS: New entry.