Plug timer wheel into scheduler
* fibers/scheduler.scm (<scheduler>): A scheduler's timers is no longer
a functional priority queue, but rather a timer wheel.  This should
decrease overhead.

There is one open question about ordering within a tick; you create a
wheel with a particular precision, and it will choose a precision at
least that precise (and no more than twice as precise).  Within a tick,
however, events are unordered.  Does it matter?  We can add a sort phase
here, perhaps when inserting the events.