Stop reading the request body in the server
Instead, the handler must read the request body if there is one to
read.

The intention here is to make it possible to process the request body
in different ways. For example, use a chunked input port if it's
chunked, or maybe just write it to disk without first reading it all
in to memory. This flexibility will make it feasible to use the fibers
web server to handle requests where the body is of an unknown size, or
large enough that you wouldn't want it in memory.