io-wakeup: Implement new I/O operations.
In retrospect, make-x/timeout? is an independent change, but it fits
and is a trivial change, so it's pointless to split it off to a new
commit.

* fibers.texi (with-read-waiting-is-failure,with-write-waiting-is-failure)
(make-read-operation,make-write-operation,accept-operation): Define
new procedures and add context.
* tests/io-wakeup.scm: Test the new procedures.
(readable/timeout?,writable/timeout?): Extract a new procedure ...
(make-x/timeout?): ..., sharing common functionality.
(accepted-socket,accept*-operation,accept*/timeout?): New stuff.
(make-listening-socket): Change from datagram sockets to stream
sockets because 'accept' is not a thing on datagram sockets and the
new tests care about that.  Also listen to the socket because a later
test needs that.
* fibers/io-wakeup.scm (with-read-waiting-is-failure,with-write-waiting-is-failure)
(make-read-operation,make-write-operation): Implement them.