texi: Remove mention of non-existent concepts in index.
This effectively merges #60, with the following changes:

The commit ‘un-comment indices’ was in the future (time-wise)
and previously (commit history) mostly duplicated in:

commit 36a6cc707921f4ab503965fefd096d80afee72f0
Author: Ludovic Courtès <ludo@gnu.org>
Date:   Mon Feb 6 14:43:20 2023 +0100

    doc: Add index.

    * fibers.texi (Index): New node.

without mentioning the previous (time-wise) PR #60.

The difference between the PR and this commit is that it appears
to merge the function and concept index.

Another difference: it's squashed.

* fibers.texi (Index): Don't mention concepts in the menu.